checkin i201103 day of month init
Exploring by writing. This constraint is not trivial to start with,
as the sequence is actually ragged. Perhaps I should use java
time's MonthDay to subtract one or whatever. Year could work this
way, but DateConstraint is too naive for use with day of month,
and possibly day of year. I can have the DateSequence be smarter,
of course, by seeing the number wrap or whatever.
Not actually finished, just committing to (possibly) change venue.